requirements.txt
aiohttp==1.0.5
async-timeout==1.2.1
certifi==2017.4.17
chardet==3.0.4
discord.py==0.16.8
future==0.16.0
idna==2.5
multidict==2.1.6
oauthlib==2.0.2
peewee==2.10.1
pkg-resources==0.0.0
python-twitter==3.3
requests==2.18.1
requests-oauthlib==0.8.0
urllib3==1.21.1
websockets==3.3
